+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Fixes:
 
 * Fixes an issue where Championships with 'Any Car Model' specified would fail to find a class for a car.
 * Fixes an issue where cars with no skins might prevent a race from starting.
-
+* Fixes an issue where Scheduled Championship Race Weekend sessions caused the calendar to error on load.
 
 v1.5.1
 ------